+++ src/components/Annotations/Group.tsx
@@ -11,7 +11,10 @@ export interface GroupProps {
 
 export default function Group(props: GroupProps) {
   const { x: oldX, y: oldY, children } = props;
-  const { x, y } = usePosition({ x: oldX, y: oldY, height: 0, width: 0 });
+  const { x, y } = usePosition({
+    x: oldX,
+    y: oldY,
+  });
 
   return <g transform={`translate(${x}, ${y})`}>{children}</g>;
 }