+++ src/components/Axis/index.tsx
@@ -4,9 +4,13 @@ import { useDispatchContext } from '../../hooks';
 import type { AxisChildProps, AxisProps } from '../../types';
 
 import BottomAxis from './BottomAxis';
+import BottomLogAxis from './BottomLogAxis';
 import LeftAxis from './LeftAxis';
+import LeftLogAxis from './LeftLogAxis';
 import RightAxis from './RightAxis';
+import RightLogAxis from './RightLogAxis';
 import TopAxis from './TopAxis';
+import TopLogAxis from './TopLogAxis';
 
 export default function Axis({
   id,
@@ -16,6 +20,7 @@ export default function Axis({
   paddingStart,
   paddingEnd,
   flip = false,
+  scale = 'linear',
 
   // children props
   label,
@@ -58,11 +63,23 @@ export default function Axis({
         paddingStart: minPadding,
         paddingEnd: maxPadding,
         flip,
+        scale,
       },
     });
 
     return () => dispatch({ type: 'removeAxis', value: { id: id || xY } });
-  }, [dispatch, xY, id, position, min, max, flip, paddingStart, paddingEnd]);
+  }, [
+    dispatch,
+    xY,
+    id,
+    position,
+    min,
+    max,
+    flip,
+    paddingStart,
+    paddingEnd,
+    scale,
+  ]);
 
   const childProps: AxisChildProps = {
     id: id || xY,
@@ -79,14 +96,22 @@ export default function Axis({
     hiddenSecondaryTicks,
   };
 
-  switch (position) {
-    case 'top':
+  switch (`${position}-${scale}`) {
+    case 'top-log':
+      return <TopLogAxis {...childProps} />;
+    case 'top-linear':
       return <TopAxis {...childProps} />;
-    case 'bottom':
+    case 'bottom-log':
+      return <BottomLogAxis {...childProps} />;
+    case 'bottom-linear':
       return <BottomAxis {...childProps} />;
-    case 'left':
+    case 'left-log':
+      return <LeftLogAxis {...childProps} />;
+    case 'left-linear':
       return <LeftAxis {...childProps} />;
-    case 'right':
+    case 'right-log':
+      return <RightLogAxis {...childProps} />;
+    case 'right-linear':
       return <RightAxis {...childProps} />;
     default:
       return null;