+++ src/hooks.ts
@@ -1,5 +1,5 @@
 import { max, min } from 'd3-array';
-import { ScaleLinear, scaleLinear } from 'd3-scale';
+import { ScaleLinear, scaleLinear, scaleLog } from 'd3-scale';
 import { createContext, useContext, useMemo } from 'react';
 
 import type {
@@ -74,13 +74,32 @@ export function useAxisContext(
       const maxPad = diff * axis.paddingEnd;
 
       const range: number[] = isHorizontal ? [0, plotWidth] : [plotHeight, 0];
-      axisContext[id] = {
-        position: axis.position,
-        scientific: diff <= 0.01 || diff >= 1000,
-        scale: scaleLinear()
-          .domain([axisMin - minPad, axisMax + maxPad])
-          .range(axis.flip ? range.reverse() : range),
-      };
+
+      switch (axis.scale) {
+        case 'log': {
+          axisContext[id] = {
+            type: axis.scale,
+            position: axis.position,
+            scientific: true,
+            scale: scaleLog()
+              .domain([axisMin - minPad, axisMax + maxPad])
+              .range(axis.flip ? range.reverse() : range),
+          };
+          break;
+        }
+        case 'linear':
+        default: {
+          axisContext[id] = {
+            type: 'linear' as const,
+            position: axis.position,
+            scientific: diff <= 0.01 || diff >= 1000,
+            scale: scaleLinear()
+              .domain([axisMin - minPad, axisMax + maxPad])
+              .range(axis.flip ? range.reverse() : range),
+          };
+          break;
+        }
+      }
     }
     return axisContext;
   }, [state, plotWidth, plotHeight]);
@@ -88,24 +107,29 @@ export function useAxisContext(
   return context;
 }
 
-type NumberString = number | string;
+type NumberOrString = number | string;
 
-interface usePositionProps {
-  x: NumberString;
-  y: NumberString;
-
-  width: NumberString;
-  height: NumberString;
+interface UsePositionConfig {
+  x: NumberOrString;
+  y: NumberOrString;
 }
 
-interface useEllipsePositionProps {
-  cx: NumberString;
-  cy: NumberString;
-  rx: NumberString;
-  ry: NumberString;
+export function usePosition(config: UsePositionConfig) {
+  const { axisContext } = usePlotContext();
+  const [xScale, yScale] = validateAxis(axisContext, 'x', 'y');
+  const { x, y } = config;
+  return {
+    x: convertValue(x, xScale),
+    y: convertValue(y, yScale),
+  };
 }
 
-export function usePosition(config: usePositionProps) {
+interface UsePositionAndSizeConfig extends UsePositionConfig {
+  width: NumberOrString;
+  height: NumberOrString;
+}
+
+export function usePositionAndSize(config: UsePositionAndSizeConfig) {
   const { axisContext } = usePlotContext();
   const [xScale, yScale] = validateAxis(axisContext, 'x', 'y');
   const { x, y, width, height } = config;
@@ -118,7 +142,14 @@ export function usePosition(config: usePositionProps) {
   };
 }
 
-export function useEllipsePosition(props: useEllipsePositionProps) {
+interface UseEllipsePositionConfig {
+  cx: NumberOrString;
+  cy: NumberOrString;
+  rx: NumberOrString;
+  ry: NumberOrString;
+}
+
+export function useEllipsePosition(props: UseEllipsePositionConfig) {
   const { axisContext } = usePlotContext();
   const [xScale, yScale] = validateAxis(axisContext, 'x', 'y');
   const { cx, cy, rx, ry } = props;