+++ src/index.ts
@@ -2,20 +2,25 @@ import Plot, { PlotProps } from './Plot';
 import PlotObject from './PlotObject';
 import * as Annotation from './components/Annotations/Annotation';
 import Axis from './components/Axis';
+import ParallelAxis from './components/Axis/ParallelAxis';
 import BarSeries from './components/BarSeries';
 import Heading from './components/Heading';
 import Legend from './components/Legend';
 import LineSeries from './components/LineSeries';
+import RangeSeries from './components/RangeSeries';
 import ScatterSeries from './components/ScatterSeries';
 import type {
   AxisProps,
+  ParallelAxisProps,
   HeadingProps,
   LegendProps,
   LineSeriesProps,
   ScatterSeriesProps,
   SeriesPointType,
+  RangeSeriesPointType,
   PlotObjectProps,
   PlotObjectType,
+  RangeSeriesProps,
 } from './types';
 
 const { Annotations } = Annotation;
@@ -25,9 +30,11 @@ export {
   Annotations,
   Plot,
   Axis,
+  ParallelAxis,
   Heading,
   Legend,
   LineSeries,
+  RangeSeries,
   ScatterSeries,
   PlotObject,
   BarSeries,
@@ -35,11 +42,14 @@ export {
 export type {
   PlotProps,
   AxisProps,
+  ParallelAxisProps,
   HeadingProps,
   LegendProps,
   LineSeriesProps,
+  RangeSeriesProps,
   ScatterSeriesProps,
   SeriesPointType,
+  RangeSeriesPointType as SeriesRangePointType,
   PlotObjectProps,
   PlotObjectType,
 };